+++ tool/src/main/java/org/kravemir/svg/labels/tool/ToolRunner.java
@@ -1,17 +1,19 @@
 package org.kravemir.svg.labels.tool;
 
+import org.kravemir.svg.labels.tool.common.CommandHelpOption;
 import picocli.CommandLine;
-
-import java.util.List;
 
 public class ToolRunner {
     public static void main(String[] args) {
-        CommandLine commandLine = new CommandLine(new TileCommand());
-        commandLine.addMixin("helperOptions", new ToolHelperOptions());
+        CommandLine commandLine = new CommandLine(new RootCommand());
 
-        commandLine.parseWithHandlers(
+        commandLine.addSubcommand("tile", new TileCommand());
+        commandLine.addSubcommand("instance", new InstanceCommand());
+
+        commandLine.addMixin("helperOptions", new CommandHelpOption());
+
+        commandLine.parseWithHandler(
                 new CommandLine.RunLast().useOut(System.out).useAnsi(CommandLine.Help.Ansi.AUTO),
-                new CommandLine.DefaultExceptionHandler<List<Object>>().useErr(System.err).useAnsi(CommandLine.Help.Ansi.AUTO),
                 args
         );
     }